# swagger-response.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parroquia API",
    "version": "1.0.0",
    "description": "Sistema de gestión parroquial - API REST completa con autenticación y gestión de usuarios",
    "contact": {
      "name": "API Support",
      "email": "support@parroquia.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "http://206.62.139.100:3000/",
      "description": "Test server"
    },
    {
      "url": "https://api.parroquia.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success",
              "error"
            ],
            "description": "Estado de la respuesta",
            "example": "success"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo",
            "example": "Operación realizada exitosamente"
          },
          "data": {
            "type": "object",
            "description": "Datos de respuesta (opcional)"
          }
        }
      },
...
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Registrar un nuevo usuario",
        "description": "Crear una nueva cuenta de usuario con verificación por email",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              },
              "example": {
                "primer_nombre": "Diego",
                "segundo_nombre": "Carlos",
                "primer_apellido": "Garcia",
                "segundo_apellido": "López",
                "correo_electronico": "diego.garcia5105@yopmail.com",
                "contrasena": "Fuerte789&",
                "telefono": "+57 300 456 7890",
                "rol": "Encuestador"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            },
                            "emailSent": {
                              "type": "boolean",
                              "description": "Si se envió el email de verificación"
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Usuario registrado exitosamente. Revisa tu email para verificar tu cuenta.",
                  "data": {
                    "user": {
                      "id": "2acc2c01-91d8-4d3a-a184-f002304620ec",
                      "correo_electronico": "carlos.martinez1024@yopmail.com",
                      "primer_nombre": "Carlos",
                      "segundo_nombre": "Alberto",
                      "primer_apellido": "Martínez",
                      "segundo_apellido": "Silva",
                      "activo": true,
                      "email_verificado": false,
                      "roles": [
                        "Encuestador"
                      ]
                    },
                    "emailSent": true
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "409": {
            "description": "Email ya registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "error",
                  "message": "El email ya está registrado",
                  "code": "EMAIL_ALREADY_EXISTS"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Iniciar sesión",
        "description": "Autenticar usuario y devolver tokens JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              },
              "example": {
                "correo_electronico": "admin@parroquia.com",
                "contrasena": "admin123"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/TokenResponse"
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Login exitoso",
                  "data": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "user": {
                      "id": "2f21d967-0ed6-4357-8876-c648550b3908",
                      "correo_electronico": "admin@parroquia.com",
                      "primer_nombre": "Admin",
                      "segundo_nombre": null,
                      "primer_apellido": "Sistema",
                      "segundo_apellido": null,
                      "activo": true,
                      "email_verificado": false,
                      "roles": [
                        "Administrador"
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "description": "Credenciales inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "error",
                  "message": "Email o contraseña incorrectos",
                  "code": "INVALID_CREDENTIALS"
                }
              }
            }
          },
          "403": {
            "description": "Cuenta inactiva o email no verificado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "error",
                  "message": "Cuenta inactiva. Contacta al administrador.",
                  "code": "ACCOUNT_INACTIVE"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/auth/refresh-token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Renovar token de acceso",
        "description": "Obtener un nuevo token de acceso usando el refresh token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenInput"
              },
              "example": {
                "refreshToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token renovado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "accessToken": {
                              "type": "string"
                            },
                            "expiresIn": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "status": "success",
                  "message": "Token renovado exitosamente",
                  "data": {
                    "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "expiresIn": 900
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "description": "Refresh token inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "status": "error",
                  "message": "Refresh token inválido o expirado",
                  "code": "INVALID_REFRESH_TOKEN"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Cerrar sesión",
        "description": "Cerrar sesión del usuario e invalidar refresh token",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logout exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                },
                "example": {
                  "status": "success",
                  "message": "Logout exitoso"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/catalog/parroquias/statistics": {
      "get": {
        "summary": "Get parroquia statistics",
        "tags": [
          "Parroquias"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}